API

Yama.SurfaceArgs — Type

Arguments for configuring loss surface computation.

Parameters

xmin::Float32, xmax::Float32, xnum::Int32: define span of the surface

and amount of point in the x direction.

ymin::Float32, ymax::Float32, ynum::Int32: define span of the surface

and amount of point in the y direction.

use_gpu::Bool: Whether to use gpu. If true then it is up to you to

transfer mini-batch in evaluation_step function to the gpu.

x_directions_file::Union{Nothing, String}: If provided, directions

for x axis will be loaded from it. Otherwise, random initialized. Should match in shape with model's weights.

y_directions_file::Union{Nothing, String}: If provided, directions

for y axis will be loaded from it. Otherwise, random initialized. Should match in shape with model's weights.

Note

If use_gpu = true then it is up to you to transfer mini-batch in evaluation_step function to the gpu.

Note

Both x- and y- directions files in should contain directions under checkpoint_weights key.

Saving weights like this in code would look like BSON.@save file checkpoint_weights.

Yama.create_surface — Method

function create_surface(
    model, dataloader::DataLoader, evaluation_step::Function, args::SurfaceArgs,
)

Create loss surface.

Parameters

model: Model to use in loss function.

dataloader::DataLoader: Dataset on which to evaluate loss function.

evaluation_step: Custom-defined function which given model, mini-batch and args, computes loss on that mini-batch.

args::SurfaceArgs: Parameters used when computing surface.

Note

If you specified use_gpu in args, then it is up to you, to transfer mini-batch in evaluation_step function to gpu.

Yama.create_trajectory — Function

function create_trajectory(
    weights::Params, checkpoints::Vector{String}, projection::Symbol = :cos,
)::Array{Float32, 2}

Create trajectory between model checkpoints. This allows for path visualization that optimizer took during training, for example.

Arguments

weights::Params: Target weights of the model. E.g. from last epoch.

checkpoints::Vector{String}: Checkpoints of the model from previous epochs.

They will be loaded in sorted order and should contain weights under checpoint_weights key.

projection::Symbol: Projection method. Either :lstsq or :cos.

Note

Checkpoint files should contain weights under checpoint_weights key.